Self Initialization (#4)
* Android Library For Initialization

* Use ctor for class loader initialization

* git remove `.idea`

* clang format

* `uint_least64_t` -> `uintptr_t`

* use explicit atomics

* Use explicit memory orderings to ensure validity of dereferenced data.

We need to change the memory orderings here to protect the values behind the pointer, the pointers themselves are atomic, and are therefore protected. Without specifying memory ordering, theoretically we could store a valid pointer to invalid data.

* Add unused attribute.

Android studio cannot detect that some things are meant to be used from the rust side, so we want to suppress the unused warning.

* typed pointers.

* get rid of `java-spaghetti` feature.

* Use `NULL` instead of 0 for uninitialized static pointers.

Co-authored-by: Suchir Kavi <suchirkavi@gmail.com>

* Don't allow dropping of global reference to application context.

Also fix class loader error message

* Use `__android_log_assert` and `__android_log_write` in ctor.

* Rename `rust_android_initialization` to `rust_android_utilities` and name the c dynlib `rust_android_utilities_setter`

---------

Co-authored-by: Suchir Kavi <suchirkavi@gmail.com>