Force LE Transport when using connect GATT
From
https://medium.com/@martijn.van.welie/making-android-ble-work-part-2-47a3cdaade07:
```
The internal implementation of the first call is that Android calls the
second one with the value TRANSPORT_AUTO for the transport parameter. If
you want to connect over BLE this is not the right value. TRANSPORT_AUTO
is for devices that support both BLE and classic Bluetooth. It means you
have ‘no preference’ for the type of connection that is made and want
Android to choose. It is totally unclear how Android chooses so this may
lead to rather unpredictable results and many people have reported
issues. That is why you should always use the second version and pass
TRANSPORT_LE for the transport parameter
```