Don't Manually Connect within automatic connection duration.
When you call `BluetoothDevice::connectGatt` Android will automatically
try to make the bluetooth connection. The method by which it does so is
controlled by the confusingly named `autoConnect` parameter, but it will
always attempt to make the connection. If `autoConnect` is set to false
(what we use in `bluedroid`), then this connection will have a 30s
timeout. It seems that some devices (e.x. Pixel 6) fail the connection
if you call `BluetoothGatt::connect` while the automatic connection
attempt is still pending. In this case they return the helpfully
unspecific error 133. This PR modifies our `connect` function to only
trigger a `BluetoothGatt::connect` if that timeout has elapsed.
Otherwise it will just wait for the result of the automatic connection.

\# Testing:
Seems to have fix the connection Issues on Nathan's Pixel 6. Discovery
still works on Android 4 (Pixel 8), and my Pixel 5a.